connect finaly work

+++ broker/mqttServer.h
@@ -5,6 +5,7 @@
 #include <threads.h>
 #include <sys/socket.h>
 #include <netdb.h>
+#include <arpa/inet.h>
 #include <ctype.h>
 
 #define MY_PORT "1883"
@@ -49,8 +50,8 @@
 /*******************************************/
 
 typedef struct {
-    int8_t messageType;
-    int16_t remainingLenght;
+    uint8_t messageType;
+    uint16_t remainingLenght;
 } fixedHeader;
 
 /*******************************************/
@@ -106,15 +107,15 @@ void handleFixedHeader(char *args);
 
 typedef struct {
     // 4
-    int16_t nameLenght;
+    uint16_t nameLenght;
     // "MQTT"
     char name[4];
     // 4 for 3.1 | 5 for 5
-    int8_t version;
+    uint8_t version;
     // type of connection
-    int8_t connectFlags;
+    uint8_t connectFlags;
     // the time in seconds between each user mqtt packet transmition
-    int16_t keepAlive;
+    uint16_t keepAlive;
 } connectVariableHeader;
 
 /*******************************************/
@@ -143,20 +144,20 @@ typedef struct {
 /*******************************************/
 
 typedef struct {
-    int16_t clientIDSize;
+    uint16_t clientIDSize;
     char* clientID;
-    int16_t willTopicSize;
+    uint16_t willTopicSize;
     char* willTopic;
-    int16_t willMessageSize;
+    uint16_t willMessageSize;
     char* willMessage;
-    int16_t userNameSize;
+    uint16_t userNameSize;
     char* userName;
-    int16_t passWordSize;
+    uint16_t passWordSize;
     char* passWord;
 } connectPayload;
 
-void handleConnect(char* args, int16_t offset);
-connectPayload readConnectPayload(char* args, int16_t offset);
+void handleConnect(char* args, int offset);
+connectPayload readConnectPayload(char* args, int offset);
 void freeConnectPayload(connectPayload* payload);
 
 //================================================================================================================
@@ -168,8 +169,8 @@ void freeConnectPayload(connectPayload* payload);
 /*******************************************/
 
 typedef struct {
-    int8_t flags;
-    int8_t returnCode;
+    uint8_t flags;
+    uint8_t returnCode;
 } connackVariableHeader;
 
 /*******************************************/
@@ -194,9 +195,9 @@ typedef struct {
 /*******************************************/
 
 typedef struct {
-    int16_t topicSize;
+    uint16_t topicSize;
     char* topic;
-    int16_t identifier;
+    uint16_t identifier;
 } publishVariableHeader;
 
 
@@ -207,11 +208,11 @@ typedef struct {
 /*******************************************/
 
 typedef struct {
-    int16_t payloadSize;
+    uint16_t payloadSize;
     char* data;
 } publishPayload;
 
-void handlePublish(char* args, int16_t offset);
+void handlePublish(char* args, int offset);
 
 //================================================================================================================
 
@@ -222,7 +223,7 @@ void handlePublish(char* args, int16_t offset);
 /*******************************************/
 
 typedef struct {
-    int16_t identifier;
+    uint16_t identifier;
 } pubackVariableHeader;
 
 //================================================================================================================
@@ -234,7 +235,7 @@ typedef struct {
 /*******************************************/
 
 typedef struct {
-    int16_t identifier;
+    uint16_t identifier;
 } subscribeVariableHeader;
 
 /*******************************************/
@@ -244,12 +245,12 @@ typedef struct {
 /*******************************************/
 
 typedef struct {
-    int16_t payloadSize;
+    uint16_t payloadSize;
     char* topic;
-    int8_t qos;
+    uint8_t qos;
 } subscribePayload;
 
-void handleSubscribe(char* args, int16_t offset);
+void handleSubscribe(char* args, int offset);
 
 //================================================================================================================
 
@@ -260,7 +261,7 @@ void handleSubscribe(char* args, int16_t offset);
 /*******************************************/
 
 typedef struct {
-    int16_t identifier;
+    uint16_t identifier;
 } subackbeVariableHeader;
 
 //================================================================================================================
